# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI Train API.",
        "title": "AI Train API",
        "contact": {},
        "license": {},
        "version": "0.2"
    },
    "host": "localhost:38080",
    "basePath": "/",
    "paths": {
        "/v1/course/create/": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add new course information into database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "Add new course information",
                "parameters": [
                    {
                        "description": "course information",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.AddCourse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericOKResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/course/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "All associated job, Deployment and svc in kubernetes are also deleted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "Delete course information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "course uuid, eg: 131ba8a9-b60b-44f9-83b5-46590f756f41",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericOKResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/course/get/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get one courses information by course id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "Get one courses information by course id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "course uuid, eg: 131ba8a9-b60b-44f9-83b5-46590f756f41",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GetCourseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/course/level/{level}": {
            "get": {
                "description": "List basic or advance courses information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "List basic or advance courses information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "basic or advance",
                        "name": "level",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ListCourseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/course/list": {
            "get": {
                "description": "get all course information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "List all course information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ListCourseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List someone's all courses information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "List someone's all courses information",
                "parameters": [
                    {
                        "description": "search user course",
                        "name": "list_user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.OauthUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ListCourseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/course/search": {
            "post": {
                "description": "Search course name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "Search course name",
                "parameters": [
                    {
                        "description": "search keyword",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.Search"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ListCourseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/course/update/": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update course information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "Update course information",
                "parameters": [
                    {
                        "description": "new course information",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.UpdateCourse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericOKResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "docs.AddCourse": {
            "type": "object",
            "properties": {
                "datasets": {
                    "type": "array",
                    "format": "string",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "cifar-10",
                        "caltech256"
                    ]
                },
                "gpu": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "image": {
                    "type": "string",
                    "format": "string",
                    "example": "nginx:1.7.9"
                },
                "introduction": {
                    "type": "string",
                    "format": "string",
                    "example": "課程說明"
                },
                "level": {
                    "type": "string",
                    "format": "string",
                    "example": "basic"
                },
                "name": {
                    "type": "string",
                    "format": "string",
                    "example": "jimmy的課"
                },
                "user": {
                    "type": "string",
                    "example": "ogre@gamil.com"
                }
            }
        },
...
        "docs.GenericErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "format": "bool",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "example": "error response message"
                }
            }
        },
        "docs.GenericOKResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "format": "bool",
                    "example": false
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "example": "response message"
                }
            }
        },
        "docs.GetCourseResponse": {
            "type": "object",
            "properties": {
                "course": {
                    "type": "object",
                    "$ref": "#/definitions/docs.Course"
                },
                "error": {
                    "type": "boolean",
                    "format": "bool",
                    "example": false
                }
            }
        },
...
        "docs.ListCourseResponse": {
            "type": "object",
            "properties": {
                "courses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/docs.Course"
                    }
                },
                "error": {
                    "type": "boolean",
                    "format": "bool",
                    "example": false
                }
            }
        },
...
        "docs.OauthUser": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "string",
                    "example": "ogre@gamil.com"
                }
            }
        },
...
        "docs.Search": {
            "type": "object",
            "properties": {
                "query": {
                    "type": "string",
                    "example": "course keyword"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}